% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_genomes.R
\name{blast_genomes}
\alias{blast_genomes}
\title{Perfrom BLAST Searches Against a Set of Genomes}
\usage{
blast_genomes(query, subject_genomes, blast_type = "blastn",
  blast_output_path = "blast_output", min_alig_length = 50, ...)
}
\arguments{
\item{query}{path to input file in fasta format.}

\item{subject_genomes}{a vector containing file paths to the reference genomes that shall be queried (e.g. file paths returned by \code{\link[biomartr]{meta.retrieval}}).}

\item{blast_type}{specification of the BLAST type shall be used to perform BLAST searches between query and reference.
Available options are:
\itemize{
\item \code{blast_type = "blastn"} for  nucleotide to nucleotide BLAST searches (see \code{\link{blast_nucleotide_to_nucleotide}} for details)
\item \code{blast_type = "tblastn"} for protein to nucleotide BLAST searches (see \code{\link{blast_protein_to_nucleotide}} for details)
}}

\item{blast_output_path}{a path to a folder that will be created to store BLAST output tables for each individual query-genome search.}

\item{min_alig_length}{minimum alignment length that shall be retained in the result dataset. All hit alignments with smaller
hit alignment length will be removed automatically.}

\item{\dots}{additional arguments passed to \code{\link{blast_nucleotide_to_nucleotide}} or \code{\link{blast_protein_to_nucleotide}}.}
}
\description{
This function takes a fasta file containing query sequences
as input and performs BLAST searches of these query sequences against
a set of reference genomes to retrieve corresponding hits in diverse genomes.
}
\details{
The \code{blast_genomes} function enables users to BLAST specific query sequences against a set of reference genomes
and retrieve the corresponding BLAST output.
}
\author{
Hajk-Georg Drost
}
