% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_nucleotide_to_genomes.R
\name{blast_nucleotide_to_genomes}
\alias{blast_nucleotide_to_genomes}
\title{Perfrom BLAST Searches Against a Set of Genomes}
\usage{
blast_nucleotide_to_genomes(
  query,
  subject_genomes,
  blast_type = "blastn",
  blast_output_path = "blast_output",
  min_alig_length = 30,
  evalue = 1e-05,
  max.target.seqs = 5000,
  update = FALSE,
  ...
)
}
\arguments{
\item{query}{path to input file in fasta format.}

\item{subject_genomes}{a vector containing file paths to the reference genomes that shall be queried (e.g. file paths returned by \code{\link[biomartr]{meta.retrieval}}).}

\item{blast_type}{specification of the BLAST type shall be used to perform BLAST searches between query and reference.
Available options are:
\itemize{
\item \code{blast_type = "blastn"} for  nucleotide to nucleotide BLAST searches (see \code{\link{blast_nucleotide_to_nucleotide}} for details)
\item \code{blast_type = "tblastn"} for protein to nucleotide BLAST searches (see \code{\link{blast_protein_to_nucleotide}} for details)
}}

\item{blast_output_path}{a path to a folder that will be created to store BLAST output tables for each individual query-genome search.}

\item{min_alig_length}{minimum alignment length that shall be retained in the result dataset. All hit alignments with smaller
hit alignment length will be removed automatically.}

\item{evalue}{minimum expectation value (E) threshold for retaining hits (default: evalue = 0.00001).}

\item{max.target.seqs}{maximum number of aligned sequences that shall be retained. Please be aware that \code{max.target.seqs} selects best hits based on the database entry and not by the best e-value. See details here: https://academic.oup.com/bioinformatics/advance-article/doi/10.1093/bioinformatics/bty833/5106166 .}

\item{update}{a logical value indicating whether or not pre-computed BLAST tables should be removed and re-computed (\code{update = TRUE})
or imported from existing file (\code{update = FALSE}) (Default).}

\item{\dots}{additional arguments passed to \code{\link{blast_nucleotide_to_nucleotide}} or \code{\link{blast_protein_to_nucleotide}}.}
}
\description{
This function takes a fasta file containing query sequences
as input and performs BLAST searches of these query sequences against
a set of reference genomes to retrieve corresponding hits in diverse genomes.
}
\details{
The \code{blast_nucleotide_to_genomes} function enables users to BLAST specific query sequences against a set of reference genomes
and retrieve the corresponding BLAST output.
}
\author{
Hajk-Georg Drost
}
