% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_nucleotide_to_nucleotide.R
\name{blast_nucleotide_to_nucleotide}
\alias{blast_nucleotide_to_nucleotide}
\title{Perform Nucleotide to Nucleotide BLAST Searches (BLASTN)}
\usage{
blast_nucleotide_to_nucleotide(query, subject, strand = "both",
  output.path = NULL, is.subject.db = FALSE, task = "megablast",
  import = FALSE, postgres.user = NULL, evalue = 0.001,
  out.format = "csv", cores = 1, max.target.seqs = 500,
  db.soft.mask = FALSE, db.hard.mask = FALSE, blast.path = NULL)
}
\arguments{
\item{query}{path to input file in fasta format.}

\item{subject}{path to subject file in fasta format or blast-able database.}

\item{strand}{Query DNA strand(s) to search against database/subject.
Options are:
\itemize{
\item \code{strand = "both"} : query against both DNA strands.
\item \code{strand = "minus"} : query against minus DNA strand.
\item \code{strand = "plus"} : query against plus DNA strand.
}}

\item{output.path}{path to folder at which BLAST output table shall be stored. 
Default is \code{output.path = NULL} (hence \code{getwd()} is used).}

\item{is.subject.db}{logical specifying whether or not the \code{subject} file is a file in fasta format (\code{is.subject.db = FALSE}; default)
or a blast-able database that was formatted with \code{makeblastdb} (\code{is.subject.db = TRUE}).}

\item{task}{nucleotide search task option. Options are:
\itemize{
\item \code{task = "blastn"} : Standard nucleotide-nucleotide comparisons (default) - Traditional BLASTN requiring an exact match of 11.
\item \code{task = "blastn-short"} : Optimized nucleotide-nucleotide comparisons for query sequences shorter than 50 nucleotides.
\item \code{task = "dc-megablast"} : Discontiguous megablast used to find more distant (e.g., interspecies) sequences.
\item \code{task = "megablast"} : Traditional megablast used to find very similar (e.g., intraspecies or closely related species) sequences.
\item \code{task = "rmblastn"}
}}

\item{import}{shall output of the nucleotide BLAST search be directly imported via \code{\link{read_blast}}? Default is \code{import = FALSE}.
In case \code{import = TRUE} only the following output formats will be imported:
\itemize{
 \item \code{out.format = "xml"} : XML
 \item \code{out.format = "tab"} : Tabular separated file
 \item \code{out.format = "csv"} : Comma-separated values
 }}

\item{postgres.user}{when \code{import = TRUE} and \code{out.format = "tab"} is selected, the BLAST output is imported and stored in a 
PostgresSQL database. In that case, users need to have PostgresSQL installed and initialized on their system. 
Please consult the Installation Vignette for details.}

\item{evalue}{Expectation value (E) threshold for saving hits (default: \code{evalue = 0.001}).}

\item{out.format}{a character string specifying the format of the file in which the BLAST results shall be stored.
Available options are:
 \itemize{
 \item \code{out.format = "pair"} : Pairwise
 \item \code{out.format = "qa.ident"} : Query-anchored showing identities
 \item \code{out.format = "qa.nonident"} : Query-anchored no identities
 \item \code{out.format = "fq.ident"} : Flat query-anchored showing identities
 \item \code{out.format = "fq.nonident"} : Flat query-anchored no identities
 \item \code{out.format = "xml"} : XML
 \item \code{out.format = "tab"} : Tabular separated file
 \item \code{out.format = "tab.comment"} : Tabular separated file with comment lines
 \item \code{out.format = "ASN.1.text"} : Seqalign (Text ASN.1)
 \item \code{out.format = "ASN.1.binary"} : Seqalign (Binary ASN.1)
 \item \code{out.format = "csv"} : Comma-separated values
 \item \code{out.format = "ASN.1"} : BLAST archive (ASN.1)
 \item \code{out.format = "json.seq.aln"} : Seqalign (JSON)
 \item \code{out.format = "json.blast.multi"} : Multiple-file BLAST JSON
 \item \code{out.format = "xml2.blast.multi"} : Multiple-file BLAST XML2
 \item \code{out.format = "json.blast.single"} : Single-file BLAST JSON
 \item \code{out.format = "xml2.blast.single"} : Single-file BLAST XML2
 \item \code{out.format = "SAM"} : Sequence Alignment/Map (SAM)
 \item \code{out.format = "report"} : Organism Report
 }}

\item{cores}{number of cores for parallel BLAST searches.}

\item{max.target.seqs}{maximum number of aligned sequences that shall be kept.}

\item{db.soft.mask}{shall low complexity regions be soft masked? Default is \code{db.soft.mask = FALSE}.}

\item{db.hard.mask}{shall low complexity regions be hard masked? Default is \code{db.hard.mask = FALSE}.}

\item{blast.path}{path to BLAST executables.}
}
\description{
Run nucleotide to nucleotide BLAST of reference sequences
against a blast-able database or fasta file.
}
\examples{
\dontrun{
blast_nucleotide_to_nucleotide(
                 query   = system.file('seqs/qry_nn.fa', package = 'metablastr'),
                 subject = system.file('seqs/sbj_nn.fa', package = 'metablastr'),
                 output.path = tempdir(),
                 import  = TRUE)
}

}
\seealso{
\code{\link{blast_protein_to_protein}}, \code{\link{blast_nucleotide_to_protein}}
}
\author{
Hajk-Georg Drost
}
