% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_p.R
\name{blast_p}
\alias{blast_p}
\title{Perform Protein to Protein BLAST Searches}
\usage{
blast_p(query, subject, out.format = "tab", cores = 1, blast.path = NULL)
}
\arguments{
\item{query}{path to input file in fasta format.}

\item{subject}{path to subject file in fasta format or blast-able database.}

\item{out.format}{a character string specifying the format of the file in which the BLAST results shall be stored.
Available options are:
 \itemize{
 \item \code{out.format = "pair"} : Pairwise
 \item \code{out.format = "qa.ident"} : Query-anchored showing identities
 \item \code{out.format = "qa.nonident"} : Query-anchored no identities
 \item \code{out.format = "fq.ident"} : Flat query-anchored showing identities
 \item \code{out.format = "fq.nonident"} : Flat query-anchored no identities
 \item \code{out.format = "xml"} : XML
 \item \code{out.format = "tab"} : Tabular separated file
 \item \code{out.format = "tab.comment"} : Tabular separated file with comment lines
 \item \code{out.format = "ASN.1.text"} : Seqalign (Text ASN.1)
 \item \code{out.format = "ASN.1.binary"} : Seqalign (Binary ASN.1)
 \item \code{out.format = "csv"} : Comma-separated values
 \item \code{out.format = "ASN.1"} : BLAST archive (ASN.1)
 \item \code{out.format = "json.seq.aln"} : Seqalign (JSON)
 \item \code{out.format = "json.blast.multi"} : Multiple-file BLAST JSON
 \item \code{out.format = "xml2.blast.multi"} : Multiple-file BLAST XML2
 \item \code{out.format = "json.blast.single"} : Single-file BLAST JSON
 \item \code{out.format = "xml2.blast.single"} : Single-file BLAST XML2
 \item \code{out.format = "report"} : Organism Report
 },}
}
\description{
Run protein to protein BLAST of reference sequences
against a blast-able database.
}
\author{
Hajk-Georg Drost
}

