% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_protein_to_nr_database.R
\name{blast_protein_to_nr_database}
\alias{blast_protein_to_nr_database}
\title{Perform Protein to Protein BLAST Searches Against the NCBI NR Database}
\usage{
blast_protein_to_nr_database(
  query,
  nr.database,
  nr.needs.formatting = FALSE,
  output.path = NULL,
  task = "blastp",
  db.import = FALSE,
  postgres.user = NULL,
  evalue = 1e-05,
  out.format = "csv",
  cores = 1,
  max.target.seqs = 65200,
  db.soft.mask = TRUE,
  db.hard.mask = TRUE,
  blast.path = NULL
)
}
\arguments{
\item{query}{path to input file in fasta format.}

\item{nr.database}{path to local NCBI NR database.}

\item{nr.needs.formatting}{a logical value indicating whether or not the local database specified in \code{nr.database} is already pre-formatted (\code{nr.needs.formatting = FALSE}; default) or whether the local NCBI NR databasse still requires formatting (\code{nr.needs.formatting = TRUE}).}

\item{output.path}{path to folder at which BLAST output table shall be stored. 
Default is \code{output.path = NULL} (hence \code{getwd()} is used).}

\item{task}{protein search task option. Options are:
\itemize{
\item \code{task = "blastp"} : Standard protein-protein comparisons (default).
\item \code{task = "blast-fast"} : Improved BLAST searches using longer words for protein seeding.
\item \code{task = "blastp-short"} : Optimized protein-protein comparisons for query sequences shorter than 30 residues.
}}

\item{db.import}{shall the BLAST output be stored in a PostgresSQL database and shall a connection be established to this database? Default is \code{db.import = FALSE}.
In case users wish to to only generate a BLAST output file without importing it to the current R session they can specify \code{db.import = NULL}.}

\item{postgres.user}{when \code{db.import = TRUE} and \code{out.format = "postgres"} is selected, the BLAST output is imported and stored in a 
PostgresSQL database. In that case, users need to have PostgresSQL installed and initialized on their system. 
Please consult the Installation Vignette for details.}

\item{evalue}{Expectation value (E) threshold for saving hits (default: \code{evalue = 0.001}).}

\item{out.format}{a character string specifying the format of the file in which the BLAST results shall be stored.
Available options are:
 \itemize{
 \item \code{out.format = "pair"} : Pairwise
 \item \code{out.format = "qa.ident"} : Query-anchored showing identities
 \item \code{out.format = "qa.nonident"} : Query-anchored no identities
 \item \code{out.format = "fq.ident"} : Flat query-anchored showing identities
 \item \code{out.format = "fq.nonident"} : Flat query-anchored no identities
 \item \code{out.format = "xml"} : XML
 \item \code{out.format = "tab"} : Tabular separated file
 \item \code{out.format = "tab.comment"} : Tabular separated file with comment lines
 \item \code{out.format = "ASN.1.text"} : Seqalign (Text ASN.1)
 \item \code{out.format = "ASN.1.binary"} : Seqalign (Binary ASN.1)
 \item \code{out.format = "csv"} : Comma-separated values
 \item \code{out.format = "ASN.1"} : BLAST archive (ASN.1)
 \item \code{out.format = "json.seq.aln"} : Seqalign (JSON)
 \item \code{out.format = "json.blast.multi"} : Multiple-file BLAST JSON
 \item \code{out.format = "xml2.blast.multi"} : Multiple-file BLAST XML2
 \item \code{out.format = "json.blast.single"} : Single-file BLAST JSON
 \item \code{out.format = "xml2.blast.single"} : Single-file BLAST XML2
 \item \code{out.format = "report"} : Organism Report
 }}

\item{cores}{number of cores for parallel BLAST searches.}

\item{max.target.seqs}{maximum number of aligned sequences that shall be retained. Please be aware that \code{max.target.seqs} selects best hits based on the database entry and not by the best e-value. See details here: https://academic.oup.com/bioinformatics/advance-article/doi/10.1093/bioinformatics/bty833/5106166 .}

\item{db.soft.mask}{shall low complexity regions be soft masked? Default is \code{db.soft.mask = FALSE}.}

\item{db.hard.mask}{shall low complexity regions be hard masked? Default is \code{db.hard.mask = FALSE}.}

\item{blast.path}{path to BLAST executables.}
}
\description{
Run protein to protein BLAST of reference sequences
against the NCBI non-redundant protein sequence Database.
}
\examples{
\dontrun{
blast_example <- blast_protein_to_nr_database(
              query   = system.file('seqs/qry_aa.fa', package = 'metablastr'),
              nr.database = "nr",
              output.path = tempdir(),
              db.import  = FALSE)
              
# look at BLAST results
blast_example
}

}
\seealso{
\code{\link{blast_nucleotide_to_nucleotide}}, \code{\link{blast_nucleotide_to_protein}},
\code{\link{blast_protein_to_nucleotide}}, \code{\link{blast_best_hit}}
}
\author{
Hajk-Georg Drost
}
