% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_homologs_cds_to_cds.R
\name{detect_homologs_cds_to_cds}
\alias{detect_homologs_cds_to_cds}
\title{Perform CDS to CDS BLAST Searches against a set of CDS files}
\usage{
detect_homologs_cds_to_cds(query, subject_cds, task = "blastn",
  blast_output_path = "blast_output", min_alig_length = 60,
  evalue = 1e-05, max.target.seqs = 5000, cores = 1, update = FALSE,
  ...)
}
\arguments{
\item{query}{path to input file in fasta format.}

\item{subject_cds}{a character vector containing paths to subject files in fasta format.}

\item{task}{nucleotide search task option. Options are:
\itemize{
\item \code{task = "blastn"} : Standard nucleotide-nucleotide comparisons (default) - Traditional BLASTN requiring an exact match of 11.
\item \code{task = "blastn-short"} : Optimized nucleotide-nucleotide comparisons for query sequences shorter than 50 nucleotides.
\item \code{task = "dc-megablast"} : Discontiguous megablast used to find more distant (e.g., interspecies) sequences.
\item \code{task = "megablast"} : Traditional megablast used to find very similar (e.g., intraspecies or closely related species) sequences.
\item \code{task = "rmblastn"}
}}

\item{blast_output_path}{a path to a folder that will be created to store BLAST output tables for each individual query-cds search.}

\item{min_alig_length}{minimum alignment length that shall be retained in the result dataset. All hit alignments with smaller hit alignment length will be removed automatically.}

\item{evalue}{Expectation value (E) threshold for saving hits (default: evalue = 1E-5).}

\item{max.target.seqs}{maximum number of aligned sequences that shall be kept. Default is \code{max.target.seqs = 500}.}

\item{cores}{number of cores for parallel BLAST searches.}

\item{update}{a logical value indicating whether or not pre-computed BLAST tables should be removed and re-computed (\code{update = TRUE}) or imported from existing file (\code{update = FALSE}) (Default).}

\item{\dots}{additional arguments passed to \code{\link{blast_nucleotide_to_nucleotide}}.}
}
\description{
Run cds to cds BLAST searches to detect homologous cds sequences
in a set of subject cds files.
}
\author{
Hajk-Georg Drost
}
