% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_hit_seqs_from_genomes.R
\name{extract_hit_seqs_from_genomes}
\alias{extract_hit_seqs_from_genomes}
\title{Extract sequences of BLAST hits in respective genomes and store it as \code{fasta} file(s)}
\usage{
extract_hit_seqs_from_genomes(blast_tbl, subject_genomes,
  file_name = NULL, separated_by_genome = FALSE, update = TRUE,
  path = NULL)
}
\arguments{
\item{blast_tbl}{a BLAST table generated with \code{\link{blast_genomes}}.}

\item{subject_genomes}{a vector containing file paths to the reference genomes that shall be queried (e.g. file paths returned by \code{\link[biomartr]{meta.retrieval}}).}

\item{file_name}{name of the fasta file that stores the BLAST hit sequences. This name will only be used when \code{separated_by_genome = FALSE}.}

\item{separated_by_genome}{a logical value indicating whether or not hit sequences from different genomes should be stored in the same
output \code{fasta} file \code{separated_by_genome = FALSE} (default) or in separate \code{fasta} files \code{separated_by_genome = TRUE}.}

\item{update}{shall an existing \code{file_name} file be overwritten (\code{update = TRUE}; Default) or shall blast hit sequences be appended to the existing file (\code{update = FALSE})?}

\item{path}{a folder path in which corresponding \code{fasta} output files shall be stored.}
}
\description{
In some cases, users may wish to extract the sequences of the respective blast hit
stored in a \code{blast_tbl}. This function enables to quickly extract such sequences and store
them in one general or multiple genome specific \code{fasta} file(s).
}
\author{
Hajk-Georg Drost
}
