% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_promotor_seqs_from_genome.R
\name{extract_promotor_seqs_from_genome}
\alias{extract_promotor_seqs_from_genome}
\title{Retrieve all upstream promotor sequences from a genome}
\usage{
extract_promotor_seqs_from_genome(annotation_file, genome_file,
  promotor_length = 500, annotation_format = "gtf", file_name = NULL,
  path = NULL, update = TRUE)
}
\arguments{
\item{annotation_file}{file path to the annotation file of the genome assembly
in \code{gtf} or \code{gff} format.}

\item{genome_file}{file path to the genome assembly file.}

\item{promotor_length}{width of upstream promotors. This is -\code{promotor_width} bp from the transcription start site (TSS) of the gene.}

\item{annotation_format}{format of the annotation file. Options are:
\itemize{
\item \code{annotation_format = "gtf"}
\item \code{annotation_format = "gff"}
}}

\item{file_name}{file path to the output file storing the promotor sequences.}

\item{path}{a file path to an output folder storing the promotor sequences.}

\item{update}{shall previously generated promotor sequences be overwritten when again generated for the same genome assembly?}
}
\description{
Given a genome assembly file and an corresponding annotation 
file users can retrieve all upstream promotor sequences of all genes from a genome.
}
\author{
Hajk-Georg Drost
}
