% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_random_seqs_from_multiple_genomes.R
\name{extract_random_seqs_from_multiple_genomes}
\alias{extract_random_seqs_from_multiple_genomes}
\title{Extract random loci from a set of genomes}
\usage{
extract_random_seqs_from_multiple_genomes(
  sample_size,
  replace = TRUE,
  prob = NULL,
  interval_width,
  subject_genomes,
  file_name = NULL,
  separated_by_genome = FALSE,
  update = TRUE,
  path = NULL
)
}
\arguments{
\item{sample_size}{a non-negative integer giving the number of loci that shall be sampled.}

\item{replace}{logical value indicating whether sampling should be with replacement. Default: \code{replace = TRUE}.}

\item{prob}{a vector of probability weights for obtaining the elements of the vector being sampled. Default is \code{prob = NULL}.}

\item{interval_width}{the length of the locus that shall be sampled.}

\item{subject_genomes}{a vector containing file paths to the reference genomes that shall be queried (e.g. file paths returned by \code{\link[biomartr]{meta.retrieval}}).}

\item{file_name}{name of the fasta file that stores the BLAST hit sequences. This name will only be used when \code{separated_by_genome = FALSE}.}

\item{separated_by_genome}{a logical value indicating whether or not hit sequences from different genomes should be stored in the same
output \code{fasta} file \code{separated_by_genome = FALSE} (default) or in separate \code{fasta} files \code{separated_by_genome = TRUE}.}

\item{update}{shall an existing \code{file_name} file be overwritten (\code{update = TRUE}; Default) or shall blast hit sequences be appended to the existing file (\code{update = FALSE})?}

\item{path}{a folder path in which corresponding \code{fasta} output files shall be stored.}
}
\description{
In some cases, users may wish to extract sequences from randomly sampled loci of a particular length from a set of genomes.
This function allows users to specify a number of sequences
of a specified length that shall be randomly sampled from the genome.
The sampling rule is as follows:
For each locus independently sample:
\itemize{
\item 1) choose randomly (equal probability: see \code{\link{sample.int}} for details) from which of the given chromosomes the locus shall be sampled (\code{replace = TRUE}).
\item 2) choose randomly (equal probability: see \code{\link{sample.int}} for details) from which strand (plus or minus) the locus shall be sampled (\code{replace = TRUE}).
\item 3) randomly choose (equal probability: see \code{\link{sample.int}} the starting position of the locus in the sampled chromosome and strand (\code{replace = TRUE}).
}
}
\author{
Hajk-Georg Drost
}
