% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_upstream_promotor_seqs.R
\name{extract_upstream_promotor_seqs}
\alias{extract_upstream_promotor_seqs}
\title{Retrieve all upstream promotor sequences from a genome}
\usage{
extract_upstream_promotor_seqs(organism, genome_file, annotation_file,
  annotation_format, file_name = NULL, promotor_width,
  replaceUnstranded = "+")
}
\arguments{
\item{organism}{a character string specifying the scientific name of the organism.}

\item{genome_file}{file path to the genome assembly file.}

\item{annotation_file}{file path to the annotation file of the genome assembly
in \code{gtf} format.}

\item{annotation_format}{format of the annotation file. Options are:
\itemize{
\item \code{annotation_format = "gtf"}
\item \code{annotation_format = "gff"}
\item \code{annotation_format = "gff3"}
}}

\item{file_name}{file path to the output file storing the promotor sequences.}

\item{promotor_width}{width of upstream promotors. This is -\code{promotor_width} bp from the
transcription start site (TSS) of the gene.}

\item{replaceUnstranded}{logical value indicating whether or not unstranded sequences shall receive a default strand. Default is \code{replaceUnstranded = TRUE}.}
}
\description{
Given a genome assembly file and an corresponding annotation 
file users can retrieve all upstream promotor sequences of all genes from a genome.
}
\details{
This function extracts genomic sequences of a specified \code{promotor_width} upstream of the transcription start sites of all genes annotated in the corresponding 
\code{annotation_file} file. The promotor sequenes are then
}
\examples{
\dontrun{
# download genome assembly of Arabidopsis lyrata
Aly_genome <- biomartr::getGenome(db = "refseq", 
                                 organism = "Arabidopsis lyrata",
                                 path = file.path("refseq", "genome"),
                                 gunzip = TRUE)
# download annotation file of genome assembly of Arabidopsis lyrata
Aly_gff <- biomartr::getGFF(db = "refseq", 
                           organism = "Arabidopsis lyrata",
                           path = file.path("refseq", "annotation"),
                           gunzip = TRUE)
                           
# retrieve upstream promotor sequences of length 1000bp
promotor_seqs <- extract_upstream_promotor_seqs(
                               organism = "Arabidopsis lyrata",
                               genome_file = Aly_genome,
                               annotation_file = Aly_gff,
                               annotation_format = "gff",
                               promotor_width = 1000)

}
}
\author{
Hajk-Georg Drost
}
