% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_contains_aa.R
\name{file_contains_aa}
\alias{file_contains_aa}
\title{Helper function to check whether input file contains amino acids as required}
\usage{
file_contains_aa(file, input_type)
}
\arguments{
\item{file}{file path to \code{fasta} file.}

\item{input_type}{a character string specifying whether input file a \code{"query"} or \code{"subject"} file.}
}
\value{
When a file contains the correct sequence type then this function
will only execute without any output. In case a file does not contains the 
correct sequence type then an error is thrown.
}
\description{
Check whether input file contains amino acids as required.
}
\examples{
# test an aa file
file_contains_aa(file = system.file('seqs/qry_aa.fa', package = 'metablastr'), input_type = "query")
}
\author{
Hajk-Georg Drost
}
