% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_homologs_core_set_n_species.R
\name{filter_homologs_core_set_n_species}
\alias{filter_homologs_core_set_n_species}
\title{Retrieve a core set of best hits that is shared across at least N species}
\usage{
filter_homologs_core_set_n_species(blast_tbl, min_qcovhsp = 50, n_species)
}
\arguments{
\item{blast_tbl}{a BLAST table generated with \code{\link{detect_homologs_proteome_to_proteome}} or \code{\link{detect_homologs_cds_to_cds}}.}

\item{min_qcovhsp}{minimum query coverage of the hit in percent \code{0..100} that shall be retained. Default value is set to \code{min_qcovhsp = 50} (= a best hit alignment must have at least 50% query coverage).}

\item{n_species}{minimum number of species in which a query sequence needs to generates valid blast hits.}
}
\description{
This function aims to retrieve a core set of best blast hits for each query sequence that is shared across all species. In other words, only query sequences
that generated blast hits in all species in the input \code{blast_tbl} were retained.
by filtering a \code{blast_tbl} using the following criteria.
A best hit is defined as (fulfilling all three critaria):
\itemize{
\item max(alig_length): only the hit having the longest alignment length is retained.
\item \code{qcovhsp >= min_qcovhsp}: only hits that have a query coverage of at least \code{min_qcovhsp} are retained.
\item max(bit_score): only the hit having the highest bit-score is retained.
}
}
\seealso{
\code{\link{filter_best_hits}}, \code{\link{filter_homologs_core_set}}
}
\author{
Hajk-Georg Drost
}
