% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_blast_hits.R
\name{gg_blast_hits}
\alias{gg_blast_hits}
\title{Plot \code{\link{blast_genomes}} Result}
\usage{
gg_blast_hits(blast_tbl, type = "scope", scope_cutoff = 0.1, alpha = 0.7,
  scale = 4, xlab = "Query", ylab = "Density over Number of BLAST Hits",
  xticks = 5, levels = NULL)
}
\arguments{
\item{blast_tbl}{a table generated with \code{\link{blast_genomes}}.}

\item{scope_cutoff}{The scope is defined as \code{1 - (abs(q_len - alig_length) / q_len))}. The \code{scope_cutoff}
defines the minimum scope that is required to retain a BLAST hit. Default is \code{scope_cutoff = 0.1} (meaning that each BLAST hit must have at least 0.1 scope).}
}
\description{
This function generates a joyplot of the \code{\link{blast_genomes}} output.
}
\author{
Hajk-Georg Drost
}
