% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_blast_hits_summary.R
\name{gg_blast_hits_summary}
\alias{gg_blast_hits_summary}
\title{Plot \code{\link{blast_nucleotide_to_genomes}} Result}
\usage{
gg_blast_hits_summary(
  blast_tbl,
  scope_cutoff = 0.1,
  alpha = 0.7,
  scale = 4,
  query_name = "query",
  xlab = "Query",
  ylab = "Density over Number of BLAST Hits",
  title = "Total number of BLAST hits: ",
  xticks = 5,
  levels = NULL
)
}
\arguments{
\item{blast_tbl}{a BLAST table generated with \code{\link{blast_nucleotide_to_genomes}}.}

\item{scope_cutoff}{The scope is defined as \code{1 - (abs(q_len - alig_length) / q_len))}. The \code{scope_cutoff}
defines the minimum scope that is required to retain a BLAST hit. Default is \code{scope_cutoff = 0.1} (meaning that each BLAST hit must have at least 0.1 scope).}

\item{alpha}{a value passed to \code{\link[ggplot2]{aes}} specifying the transparency of distributions.}

\item{scale}{the ridges scale passed to \code{\link[ggridges]{geom_density_ridges}}.}

\item{query_name}{name of the query in the \code{blast_tbl}.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{title}{figure title which will be concatenated with the number of BLAST hits available in the inpit \code{blast_tbl}.}

\item{xticks}{number of ticks on the x-axis. Default is \code{xticks = 5}.}

\item{levels}{a character vector specifying the exact order of species names (levels) 
that is used to \code{\link{factor}} and sort species in the \code{\link[ggridges]{geom_density_ridges}} plot.}
}
\description{
This function generates a joyplot of the \code{\link{blast_nucleotide_to_genomes}} output.
}
\author{
Hajk-Georg Drost
}
