% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_detect_homologs.R
\name{gg_detect_homologs}
\alias{gg_detect_homologs}
\title{Plot \code{detect_homologs_genome_to_genome} or \code{detect_homologs_proteome_to_proteome} Result}
\usage{
gg_detect_homologs(blast_tbl, type = "perc_identity", xlab = "Species",
  ylab = "", text_size = 18, y_ticks = 8)
}
\arguments{
\item{blast_tbl}{a BLAST table generated with \code{\link{detect_homologs_proteome_to_proteome}}.}

\item{type}{the type of BLAST hit visualization. Options are:
\itemize{
\item \code{type = "perc_identity"}: visualize the alignment identity in percent for each BLAST hit.
\item \code{type = "alig_length"}: visualize the alignment length for each BLAST hit.
\item \code{type = "scope"}: visualize the length homology to the initial query for each BLAST hit.
}}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{text_size}{size of label text.}

\item{y_ticks}{number of ticks on the y-axis.}
}
\description{
This function generates a boxplot of different types of the \code{\link{detect_homologs_proteome_to_proteome}} output.
}
\author{
Hajk-Georg Drost
}
