% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_hits_boxplot.R
\name{gg_hits_boxplot}
\alias{gg_hits_boxplot}
\title{Boxplot of a \code{blast_tbl} generated with \code{detect_homologs_genome_to_genome} or \code{detect_homologs_proteome_to_proteome}}
\usage{
gg_hits_boxplot(
  blast_tbl,
  x_type = "species",
  y_type = "qcovhsp",
  xlab = x_type,
  ylab = y_type,
  title = "",
  text_size = 18,
  y_ticks = 8
)
}
\arguments{
\item{blast_tbl}{a BLAST table generated with \code{\link{detect_homologs_proteome_to_proteome}}.}

\item{x_type}{variable thah shall be visualized on the x-axis. E.g.:
\itemize{
\item \code{type = "species"}: the species in which blast hits were found (Default)
}}

\item{y_type}{variable thah shall be visualized on the y-axis. E.g.:
\itemize{
\item \code{type = "qcovhsp"}: query coverage of the respective blast hit (Default)
\item \code{type = "perc_identity"}: visualize the alignment identity in percent for each BLAST hit.
\item \code{type = "alig_length"}: visualize the alignment length for each BLAST hit.
\item \code{type = "scope"}: visualize the length homology to the initial query for each BLAST hit.
}}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{title}{plot title.}

\item{text_size}{size of label text.}

\item{y_ticks}{number of ticks on the y-axis.}
}
\description{
This function generates a boxplot of different types of the \code{\link{detect_homologs_proteome_to_proteome}} output.
}
\author{
Hajk-Georg Drost
}
