% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_homologs_one_to_one.R
\name{gg_homologs_one_to_one}
\alias{gg_homologs_one_to_one}
\title{Plot number of BLAST hits per species as barplot from a \code{tibble} generated with \code{detect_homologs_genome_to_genome} or \code{detect_homologs_proteome_to_proteome}}
\usage{
gg_homologs_one_to_one(
  blast_tbl,
  xlab = "Species",
  ylab = "Number of one-to-one homologs (including splice variants)",
  title = "One-to-one homologs",
  text_size = 18,
  y_ticks = 8
)
}
\arguments{
\item{blast_tbl}{a BLAST table generated with \code{\link{detect_homologs_proteome_to_proteome}}.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{title}{plot title.}

\item{text_size}{size of label text.}

\item{y_ticks}{number of ticks on the y-axis.}
}
\description{
This function generates a boxplot of different types of the \code{\link{detect_homologs_proteome_to_proteome}} output.
}
\author{
Hajk-Georg Drost
}
