% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_pairwise_orthologs_line.R
\name{gg_pairwise_orthologs_line}
\alias{gg_pairwise_orthologs_line}
\title{Plot number of pairwise orthologs for multiple species comparisons}
\usage{
gg_pairwise_orthologs_line(ortho_summary, xlab = "Subject Species",
  ylab = "Number of reciprocal best hit orthologs", title = "",
  vline = 7000, ymax = 24000)
}
\arguments{
\item{ortho_summary}{a \code{data.frame} storing the \code{subject_species} in the first column and the number of orthologs named \code{n_orthologs} in the second column.}

\item{xlab}{label of x-axis.}

\item{ylab}{label of y-axis.}

\item{title}{plot title.}

\item{vline}{plot a line characterizing the core set.}

\item{ymax}{max value of y-axis.}
}
\description{
Visualizes pairwise orthologs for multiple species comparisons as line graph.
}
\author{
Hajk-Georg Drost
}
