% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_species_dnds_blast_tbl.R
\name{gg_species_dnds_blast_tbl}
\alias{gg_species_dnds_blast_tbl}
\title{Plot pairwise dN, dS, or dNdS distributions of orthologous genes}
\usage{
gg_species_dnds_blast_tbl(blast_tbl, type, order, xlab = "", ylab = "",
  title = "")
}
\description{
Plot pairwise dN, dS, or dNdS distributions of orthologous genes
}
