% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_outformats.R
\name{list_outformats}
\alias{list_outformats}
\title{List available BLAST output formats}
\usage{
list_outformats()
}
\description{
Available BLAST output formats are listed as they can
be specified as \code{out.format} argument in any \code{blast_*} function.
}
\examples{
# list available BLAST output formats
list_outformats()
}
\author{
Hajk-Georg Drost
}
