% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_compare_multi_promotor_seqs.R
\name{motif_compare_multi_promotor_seqs}
\alias{motif_compare_multi_promotor_seqs}
\title{Count the number of motifs in a set of non-random versus random sequences for multiple species}
\usage{
motif_compare_multi_promotor_seqs(blast_tbl, subject_genomes,
  annotation_files, annotation_format = "gff", interval_width, motifs,
  max.mismatch = 0, min.mismatch = 0, ...)
}
\description{
Compare the number of motifs in a set of non-random versus random sequences. The resulting values can then be used to test the enrichment of certain motifs in real sequences compared to random sequences. Each enrichment analysis is performed for a set of different species/genomes.
}
\author{
Hajk-Georg Drost
}
