% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_count.R
\name{motif_count}
\alias{motif_count}
\title{Count the number of motifs in a set of sequences}
\usage{
motif_count(file, motif, max.mismatch = 0, min.mismatch = 0, ...)
}
\arguments{
\item{file}{a file path to the \code{fasta} file storing the sequences of interest.}

\item{motif}{a character string specifying the sequence motif that shall be searched.}

\item{max.mismatch}{the maximum number of mismatching letters allowed (see \code{\link[Biostrings]{matchPattern}} for details).}

\item{min.mismatch}{the minimum number of mismatching letters allowed (see \code{\link[Biostrings]{vcountPattern}} for details).}

\item{\dots}{additional arguments passed to \code{\link[Biostrings]{matchPattern}}.}
}
\description{
Detect all matches of a given motif (typically a short string) in a (typically long sequence) reference sequence or set of reference sequences.
}
\author{
Hajk-Georg Drost
}
