% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_count_matrix.R
\name{motif_count_matrix}
\alias{motif_count_matrix}
\title{Count the number of motifs in a set of sequences and generate a motif count matrix}
\usage{
motif_count_matrix(file, motifs, max.mismatch = 0, min.mismatch = 0,
  ...)
}
\arguments{
\item{file}{a file path to the \code{fasta} file storing the sequences of interest.}

\item{motifs}{a character string specifying the sequence motifs that shall be screened and counted.}

\item{max.mismatch}{the maximum number of mismatching letters allowed (see \code{\link[Biostrings]{matchPattern}} for details).}

\item{min.mismatch}{the minimum number of mismatching letters allowed (see \code{\link[Biostrings]{vcountPattern}} for details).}

\item{\dots}{additional arguments passed to \code{\link[Biostrings]{matchPattern}}.}
}
\description{
Detect all matches of a given motif (typically a short string) in a (typically long sequence) reference sequence or set of reference sequences. For each sequence and each motif a count matrix is generated and returned.
}
\author{
Hajk-Georg Drost
}
