% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment.R
\name{motif_enrichment}
\alias{motif_enrichment}
\title{Test motif enrichment in a set of non-random versus random sequences}
\usage{
motif_enrichment(real_seqs, random_seqs, motifs, test = "fisher",
  max.mismatch = 0, min.mismatch = 0, ...)
}
\arguments{
\item{real_seqs}{a file path to the \code{fasta} file storing the non-random set of sequences.}

\item{random_seqs}{a file path to the \code{fasta} file storing the random set of sequences, e.g. generated with \code{\link{extract_random_seqs_from_genome}}.}

\item{motifs}{a character vector storing a set of motifs that shall be counted within respective sequences.}

\item{test}{tests statistics and models to quantify significant motif enrichment. Options are:
\itemize{
\item \code{test = "fisher"}: Fisher's Exact Test for Count Data (see \code{link[stats]{fisher.test}} for details).
}}

\item{max.mismatch}{the maximum number of mismatching letters allowed (see \code{\link[Biostrings]{matchPattern}} for details).}

\item{min.mismatch}{the minimum number of mismatching letters allowed (see \code{\link[Biostrings]{vcountPattern}} for details).}

\item{\dots}{additional arguments passed to \code{\link[Biostrings]{matchPattern}}.}
}
\description{
Compare the number of motifs in a set of non-random versus random sequences.
The resulting values are then tested for enrichment of certain motifs in real sequences compared to random sequences. Several tests statistics and approaches are available to quantify significant motif enrichment.
}
\seealso{
\code{\link{motif_count}}, \code{\link{extract_random_seqs_from_genome}}, \code{\link{extract_hit_seqs_from_genomes}}
}
\author{
Hajk-Georg Drost
}
