% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_multi.R
\name{motif_enrichment_multi}
\alias{motif_enrichment_multi}
\title{Statistical assessment of motif enrichments in a set of non-random versus randomly sampled genomic sequences for multiple species}
\usage{
motif_enrichment_multi(blast_tbl, subject_genomes, test = "fisher",
  alternative = "two.sided", size, interval_width, motifs,
  max.mismatch = 0, min.mismatch = 0, ...)
}
\arguments{
\item{blast_tbl}{a blast_table.}

\item{subject_genomes}{a character vector storing the file paths to the subject genomes that shall be used as subject references.}

\item{test}{\itemize{
\item \code{test = "fisher"}: Fisher's Exact Test for Count Data (see \code{link[stats]{fisher.test}} for details).
}}

\item{alternative}{indicates the alternative hypothesis and must be one of \code{"two.sided"}, \code{"greater"} or \code{"less"}. You can specify just the initial letter. Only used in the 2 by 2 case.}

\item{size}{total number of sequences that shall be sampled per subject genome.}

\item{interval_width}{length of the sequence in which motifs shall be detected.}

\item{motifs}{a character vector storing (case sensitive) motif sequences for which abundance in the sampled sequences shall be assessed.}

\item{max.mismatch}{the maximum number of mismatching letters allowed (see \code{\link[Biostrings]{matchPattern}} for details).}

\item{min.mismatch}{the minimum number of mismatching letters allowed (see \code{\link[Biostrings]{vcountPattern}} for details).}

\item{...}{additional arguments passed to \code{\link{motif_enrichment}}.}
}
\description{
Compare the number of motifs in a set of non-random versus random sequences genomic sequences within a set of subject genomes. 
The resulting values are then used to statistically assess the enrichment of certain motifs in real sequences compared to randomly sampled genomic sequences.
}
\seealso{
\code{\link{motif_count}}, \code{\link{motif_compare_multi_promotor_seqs}}, \code{\link{motif_compare}}, 
\code{\link{motif_enrichment}}
}
\author{
Hajk-Georg Drost
}
