% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_multi_promotor_seqs.R
\name{motif_enrichment_multi_promotor_seqs}
\alias{motif_enrichment_multi_promotor_seqs}
\title{Statistical assessment of motif enrichments in a set of non-random versus randomly sampled gene promotor sequences for multiple species}
\usage{
motif_enrichment_multi_promotor_seqs(blast_tbl, subject_genomes,
  annotation_files, annotation_format = "gff", test = "fisher",
  alternative = "two.sided", interval_width, motifs, max.mismatch = 0,
  min.mismatch = 0, ...)
}
\arguments{
\item{blast_tbl}{a blast_table.}

\item{subject_genomes}{a character vector storing the file paths to the subject genomes that shall be used as subject references.}

\item{annotation_files}{a character vector storing the file paths to the subject annotation files in \code{.gff} format that match the subject genomes.}

\item{annotation_format}{the annotation format. Options are:
\itemize{
\item \code{annotation_format = "gff"}
}}

\item{test}{\itemize{
\item \code{test = "fisher"}: Fisher's Exact Test for Count Data (see \code{link[stats]{fisher.test}} for details).
}}

\item{alternative}{indicates the alternative hypothesis and must be one of \code{"two.sided"}, \code{"greater"} or \code{"less"}. You can specify just the initial letter. Only used in the 2 by 2 case.}

\item{interval_width}{total number of sequences that shall be sampled per subject genome.}

\item{motifs}{a character vector storing (case sensitive) motif sequences for which abundance in the sampled sequences shall be assessed.}

\item{max.mismatch}{maximum number of mismatches that are allowed between the sequence motif and the matching region in the sampled sequence.}

\item{min.mismatch}{minimum number of mismatches that are allowed between the sequence motif and the matching region in the sampled sequence.}

\item{...}{additional arguments passed to \code{\link{motif_compare}}.}
}
\description{
Compare the number of motifs in a set of non-random versus randomly sampled gene promotor sequences within a set of subject genomes. 
The resulting values are then used to statistically assess the enrichment of certain motifs in real sequences compared to randomly sampled gene promotor sequences.
}
\seealso{
\code{\link{motif_count}}, \code{\link{motif_count_multi}}, \code{\link{motif_compare_multi_promotor_seqs}}, \code{\link{motif_compare}}, 
\code{\link{motif_enrichment}}
}
\author{
Hajk-Georg Drost
}
