% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_chromosome_intervals.R
\name{sample_chromosome_intervals}
\alias{sample_chromosome_intervals}
\title{Helper function to sample random intervals of length \code{interval_width} from chromosomes}
\usage{
sample_chromosome_intervals(chr_size, interval_width, strand, size)
}
\arguments{
\item{chr_size}{an integer value defining the length of the chromosome.}

\item{interval_width}{the length of the interval that shall be sampled.}

\item{strand}{from which strand should the interval be sampled? Options
are:
\itemize{
\item \code{strand = "plus"}
\item \code{strand = "minus"}
}}

\item{size}{a non-negative integer giving the number of loci that shall be sampled.}
}
\description{
This function samples a random locus given the chromosome lenght (\code{chr_size}) 
and \code{interval_width} from a chromosome
}
\author{
Hajk-Georg Drost
}
