% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_blast_hits.R
\name{gg_blast_hits}
\alias{gg_blast_hits}
\title{Plot \code{\link{blast_nucleotide_to_genomes}} Result}
\usage{
gg_blast_hits(
  blast_tbl,
  type = "scope",
  scope_cutoff = 0.1,
  alpha = 0.7,
  scale = 4,
  xlab = "Query",
  ylab = "Density over Number of BLAST Hits",
  title = "Total number of BLAST hits: ",
  xticks = 5,
  levels = NULL,
  trim = TRUE
)
}
\arguments{
\item{blast_tbl}{a BLAST table generated with \code{\link{blast_nucleotide_to_genomes}}.}

\item{type}{the type of BLAST hit visualization. Options are:
\itemize{
\item \code{type = "scope"} 
\item \code{type = "alig_length"}
\item \code{type = "evalue"}
\item \code{type = "bit_score"}
\item \code{type = "qcovhsp"}
\item \code{type = "qcov"}
\item \code{type = "perc_identity"}
}}

\item{scope_cutoff}{The scope is defined as \code{1 - (abs(q_len - alig_length) / q_len))}. The \code{scope_cutoff}
defines the minimum scope that is required to retain a BLAST hit. Default is \code{scope_cutoff = 0.1} (meaning that each BLAST hit must have at least 0.1 scope).}

\item{alpha}{a value passed to \code{\link[ggplot2]{aes}} specifying the transparency of distributions.}

\item{scale}{the ridges scale passed to \code{\link[ggridges]{geom_density_ridges}}.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{title}{figure title which will be concatenated with the number of BLAST hits available in the inpit \code{blast_tbl}.}

\item{xticks}{number of ticks on the x-axis. Default is \code{xticks = 5}.}

\item{levels}{a character vector specifying the exact order of species names (levels) 
that is used to \code{\link{factor}} and sort species in the \code{\link[ggridges]{geom_density_ridges}} plot.}

\item{trim}{a logical value indicating whether ridges densities shall be trimmed. Default is \code{trim = TRUE}.}
}
\description{
This function generates a joyplot of the \code{\link{blast_nucleotide_to_genomes}} output.
}
\author{
Hajk-Georg Drost
}
