% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_pairwise_orthologs_line.R
\name{gg_pairwise_orthologs_line}
\alias{gg_pairwise_orthologs_line}
\title{Plot number of pairwise orthologs for multiple species comparisons}
\usage{
gg_pairwise_orthologs_line(
  ortho_summary,
  xlab = "Subject Species",
  ylab = "Number of reciprocal best hit orthologs",
  title = "",
  vline = NULL,
  ymax = NULL
)
}
\arguments{
\item{ortho_summary}{a \code{data.frame} storing the \code{subject_species} in the first column and the number of orthologs named \code{n_orthologs} in the second column.}

\item{xlab}{label of x-axis.}

\item{ylab}{label of y-axis.}

\item{title}{plot title.}

\item{vline}{plot a line characterizing the core set. Default is \code{vline = NULL} meaning that no vline is drawn.}

\item{ymax}{max value of y-axis. Default is \code{ymax = NULL} meaning that the largest value within the dataset will be used to determine y-max.}
}
\description{
Visualizes pairwise orthologs for multiple species comparisons as line graph.
}
\author{
Hajk-Georg Drost
}
