% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_blast.R
\name{read_blast}
\alias{read_blast}
\title{Import BLAST output file}
\usage{
read_blast(file, out.format, postgres.user = NULL)
}
\arguments{
\item{file}{path to BLAST output file.}

\item{out.format}{a character string specifying the output format of the BLAST output that shall be imported.
Available options are:
 \itemize{
 \item \code{out.format = "postgres"} : store BLAST output as Postgres database and generate postgres database connection.
 \item \code{out.format = "xml"} : XML
 \item \code{out.format = "tsv"} : Tabular separated file
 \item \code{out.format = "csv"} : Comma-separated values
 \item \code{out.format = "json.seq.aln"} : Seqalign (JSON)
 \item \code{out.format = "json.blast.multi"} : Multiple-file BLAST JSON
 \item \code{out.format = "xml2.blast.multi"} : Multiple-file BLAST XML2
 \item \code{out.format = "json.blast.single"} : Single-file BLAST JSON
 \item \code{out.format = "xml2.blast.single"} : Single-file BLAST XML2
 }}

\item{postgres.user}{specify username for RPostgreSQL connection.}
}
\description{
When performing BLAST searches with the \code{blast_*()} functions,
the corresponding BLAST output file can be imported into the current R session using this function.
All output formats given by BLAST are supported (see e.g. description for details).
}
\seealso{
\code{\link{blast_protein_to_protein}}, \code{\link{blast_nucleotide_to_protein}},
\code{\link{blast_nucleotide_to_nucleotide}}
}
\author{
Hajk-Georg Drost
}
