% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_chromosome_intervals.R
\name{sample_chromosome_intervals}
\alias{sample_chromosome_intervals}
\title{Helper function to sample random intervals of length \code{interval_width} from chromosomes}
\usage{
sample_chromosome_intervals(
  chr_size,
  interval_width,
  strand,
  size,
  replace = TRUE,
  prob = NULL
)
}
\arguments{
\item{chr_size}{an integer value defining the length of the chromosome.}

\item{interval_width}{the length of the interval that shall be sampled.}

\item{strand}{from which strand should the interval be sampled? Options
are:
\itemize{
\item \code{strand = "plus"}
\item \code{strand = "minus"}
}}

\item{size}{a non-negative integer giving the number of sequences that shall be sampled from random genomic loci.}

\item{replace}{logical value indicating whether sampling should be with replacement. Default: \code{replace = TRUE}.}

\item{prob}{a vector of probability weights for obtaining the elements of the vector being sampled. Default is \code{prob = NULL}.}
}
\description{
This function samples a random locus given the chromosome lenght (\code{chr_size}) 
and \code{interval_width} from a chromosome
}
\author{
Hajk-Georg Drost
}
